Add cli argument to turn off smart file checking
